Reduce time to build exportable branch
Problem: a user may SIGINT during git-publish's execution. This could
leave their git repo in a state where escape is not entirely intuitive.

To minimize the window of time in which this might happen, build the
branch and restore the old one immediately rather than trying to push
the built branch in between. Once the user's working state has been
restored then push the built branch.